+++ app/src/main/java/de/passwordvault/view/settings/activity_help/SettingsHelpActivity.java
@@ -1,11 +1,18 @@
 package de.passwordvault.view.settings.activity_help;
 
 import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.ProgressBar;
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.lifecycle.ViewModel;
+import androidx.recyclerview.widget.RecyclerView;
 import de.passwordvault.R;
-import de.passwordvault.view.settings.activity_localized_asset_viewer.LocalizedAssetViewerActivity;
+import de.passwordvault.model.rest.RestCallback;
+import de.passwordvault.model.rest.RestError;
+import de.passwordvault.model.rest.help.LocalizedHelpPage;
 import de.passwordvault.view.utils.components.PasswordVaultActivity;
 
 
@@ -13,13 +20,53 @@ import de.passwordvault.view.utils.components.PasswordVaultActivity;
  * Class implements the help activity displaying all help pages to the user.
  *
  * @author  Christian-2003
- * @version 3.5.3
+ * @version 3.7.2
  */
-public class SettingsHelpActivity extends PasswordVaultActivity<ViewModel> {
+public class SettingsHelpActivity extends PasswordVaultActivity<SettingsHelpViewModel> implements RestCallback {
 
+    /**
+     * Attribute stores the recycler view adapter for the activity.
+     */
+    private SettingsHelpRecyclerViewAdapter adapter;
+
+    /**
+     * Attribute stores the recycler view of the activity.
+     */
+    private RecyclerView recyclerView;
+
+    /**
+     * Attribute stores the progress bar indicating that data is being fetched.
+     */
+    private ProgressBar progressBar;
+
+
+    /**
+     * Constructor instantiates a new activity.
+     */
     public SettingsHelpActivity() {
-        super(null, R.layout.activity_settings_help);
+        super(SettingsHelpViewModel.class, R.layout.activity_settings_help);
     }
+
+
+    /**
+     * Method is called once the data has been fetched.
+     *
+     * @param tag   Tag used with the REST client.
+     * @param error Error generated during the call to the REST API.
+     */
+    @Override
+    public void onFetchFinished(@Nullable String tag, @NonNull RestError error) {
+        if (viewModel.isFinished()) {
+            viewModel.setError(error);
+            Log.d("REST", "Error code: " + error.ordinal());
+            runOnUiThread(() -> {
+                progressBar.setVisibility(View.GONE);
+                recyclerView.setVisibility(View.VISIBLE);
+                adapter.notifyDataSetChanged();
+            });
+        }
+    }
+
 
     /**
      * Method is called whenever the activity is created.
@@ -30,26 +77,32 @@ public class SettingsHelpActivity extends PasswordVaultActivity<ViewModel> {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        progressBar = findViewById(R.id.progress_bar);
+        progressBar.setVisibility(View.VISIBLE);
+
         //Back button:
         findViewById(R.id.button_back).setOnClickListener(view -> finish());
 
-        //Help pages:
-        findViewById(R.id.settings_help_login_container).setOnClickListener(view -> showHelpPage("configure_login.html"));
-        findViewById(R.id.settings_help_autofill_container).setOnClickListener(view -> showHelpPage("configure_autofill.html"));
-        findViewById(R.id.settings_help_backup_container).setOnClickListener(view -> showHelpPage("backup.html"));
+        //Recycler view:
+        adapter = new SettingsHelpRecyclerViewAdapter(this, viewModel);
+        adapter.setHelpPageClickListener(this::showHelpPage);
+        recyclerView = findViewById(R.id.recycler_view);
+        recyclerView.setAdapter(adapter);
+        recyclerView.setVisibility(View.GONE);
+
+        //Fetch data:
+        viewModel.fetchData(this);
     }
 
 
     /**
-     * Method hows the help page with the specified name to the user.
+     * Method hows the help page whose position is passed to the user.
      *
-     * @param name  Name of the page (e.g. "help.html") to show to the user.
+     * @param position  Position of the localized help page to show to the user.
      */
-    private void showHelpPage(String name) {
-        Intent intent = new Intent(this, LocalizedAssetViewerActivity.class);
-        intent.putExtra(LocalizedAssetViewerActivity.KEY_PAGE, name);
-        intent.putExtra(LocalizedAssetViewerActivity.KEY_FOLDER, "help");
-        startActivity(intent);
+    private void showHelpPage(int position) {
+        LocalizedHelpPage page = viewModel.getHelpPages().get(position - SettingsHelpRecyclerViewAdapter.OFFSET_HELP_PAGES);
+        openUrlInBrowserOrApp(page.getUrl());
     }
 
 }